# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "version": "1.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener todas las tareas",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Task List"
                    },
                    "404": {
                        "description": "No tasks found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nueva tarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion nueva tarea"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Task"
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "4732012736ffb7728b7e6b7949d3c140",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Task"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "3aa12cf460ff0476704f02693959d7de",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarea eliminada"
                    }
                }
            }
        },
        "/api/tasks/completed-last-week": {
            "get": {
                "tags": [
                    "completed-last-week"
                ],
                "summary": "Obtener todas las tareas completadas de la ultima semana",
                "operationId": "66d2ae6ce527d4e7dc77734fe139eb13",
                "responses": {
                    "200": {
                        "description": "Lista de tareas"
                    },
                    "404": {
                        "description": "No tasks found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "": {
                "type": "http",
                "description": "Task Api CRUD",
                "name": "Task API"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Endpoints para la gestión de tareas"
        },
        {
            "name": "completed-last-week",
            "description": "completed-last-week"
        }
    ]
}